backend/requirements/requirements.txt
Django==5.2.7
pillow==12.0.0
djangorestframework==3.16.1
